refactor: Move authentication navigation to authenticationUi
This commit moves authentication-related navigation classes and interfaces from the `feature/authentication/authenticationApi` module to the `feature/authentication/authenticationUi` module. This centralizes navigation logic within the UI module.

The following changes are included:

- Moved `AuthenticationDestination.kt`, `AuthenticationDestinations.kt`, and `AuthenticationGraph.kt` from `authenticationApi` to `authenticationUi/navigation`.
- Updated import statements in various files to reflect the new locations of these navigation components.
- Added Kotlin serialization plugin and dependency to `feature/authentication/authenticationUi/build.gradle.kts` to support serialization of navigation destinations.
- Removed `fromJsonToAuthenticationDestination` from `appNavigation/AppNavGraph.kt` as it's no longer needed in that location due to the move.
- Updated `FeaturesApiModule.kt` to use `AuthenticationDestinations` from the new location.
- Ensured `AuthenticationNavGraph.kt` and ViewModels within `authenticationUi` use the navigation components from their new location.
