--- feature/authentication/authenticationApi/src/main/java/com/feature/authentication/authenticationApi/AuthenticationGraph.kt
@@ -1,4 +1,0 @@
-package com.feature.authentication.authenticationApi
-
-
-interface AuthenticationGraph
