+++ feature/authentication/authenticationUi/build.gradle.kts
@@ -2,6 +2,7 @@ plugins {
     alias(libs.plugins.android.library)
     alias(libs.plugins.kotlin.android)
     alias(libs.plugins.kotlin.compose)
+    alias(libs.plugins.kotlin.serialization)
 }
 
 android {
@@ -88,4 +89,7 @@ dependencies {
 
     //WebView
     implementation(libs.androidx.webkit)
+
+    //kotlinx serialization
+    implementation(libs.kotlinx.serialization.json)
 }
