+++ feature/authentication/authenticationUi/src/main/java/com/feature/authentication/authenticationUi/navigation/AuthenticationDestination.kt
@@ -1,5 +1,6 @@
-package com.feature.authentication.authenticationApi
+package com.feature.authentication.authenticationUi.navigation
 
+import kotlinx.serialization.InternalSerializationApi
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.json.Json
 import kotlinx.serialization.serializer
@@ -7,7 +8,7 @@ import kotlinx.serialization.serializer
 @Serializable
 sealed interface AuthenticationDestination
 
-@OptIn(kotlinx.serialization.InternalSerializationApi::class)
+@OptIn(InternalSerializationApi::class)
  fun AuthenticationDestination.toJson(): String =
     Json.encodeToString(AuthenticationDestination::class.serializer(), this)
