+++ feature/authentication/authenticationUi/src/main/java/com/feature/authentication/authenticationUi/navigation/AuthenticationGraph.kt
@@ -1,0 +1,4 @@
+package com.feature.authentication.authenticationUi.navigation
+
+
+interface AuthenticationGraph
