+++ feature/authentication/authenticationUi/src/main/java/com/feature/authentication/authenticationUi/navigation/AuthenticationNavigationEvent.kt
@@ -1,7 +1,6 @@
 package com.feature.authentication.authenticationUi.navigation
 
 import androidx.navigation.NavOptions
-import com.feature.authentication.authenticationApi.AuthenticationDestination
 
 sealed class AuthenticationNavigationEvent {
     data class Navigate(
